Avant modification du modele Theme ( theme, theme detail)

+++ API/Modeles/ThemeWithAdress.cs
@@ -26,9 +26,9 @@ namespace API.Modeles
 
         public string TypeDePaiement { get; set; }
 
-        public int? PrixMin { get; set; }
+        public decimal? PrixMin { get; set; }
 
-        public int? PrixMax { get; set; }
+        public decimal? PrixMax { get; set; }
 
     }
 }
